Restructure loss.py into function.py
